#173 refactor basis agendas and move last work function into FileHub
