+++ src/change/test_special_func/test_add_pledge.py
@@ -1,39 +1,16 @@
+from src._road.road import get_default_real_id_roadnode as root_label
 from src.change.filehub import filehub_shop
-from src.real.admin_duty import append_changes_to_duty_file, add_pledge_change
-from src.real.examples.example_changes import sue_2atomunits_changeunit
-from src.real.examples.real_env_kit import reals_dir_setup_cleanup
+from src.change.special_func import add_pledge_change
+from src.change.examples.change_env import (
+    env_dir_setup_cleanup,
+    get_change_temp_env_dir as env_dir,
+)
 
 
-def test_append_changes_to_duty_file_AddschangesToDutyFile(
-    reals_dir_setup_cleanup,
-):
+def test_add_pledge_change_Addspledgechange(env_dir_setup_cleanup):
     # GIVEN
     sue_text = "Sue"
-    sue_filehub = filehub_shop(None, None, sue_text)
-    sue_filehub.initialize_change_duty_files()
-    sue_filehub.save_change_file(sue_2atomunits_changeunit())
-    duty_agenda = sue_filehub.get_duty_agenda()
-    print(f"{duty_agenda._real_id=}")
-    sports_text = "sports"
-    sports_road = duty_agenda.make_l1_road(sports_text)
-    knee_text = "knee"
-    knee_road = duty_agenda.make_road(sports_road, knee_text)
-    assert duty_agenda.idea_exists(sports_road) == False
-    assert duty_agenda.idea_exists(knee_road) == False
-
-    # WHEN
-    new_agenda = append_changes_to_duty_file(sue_filehub)
-
-    # THEN
-    assert new_agenda != duty_agenda
-    assert new_agenda.idea_exists(sports_road)
-    assert new_agenda.idea_exists(knee_road)
-
-
-def test_add_pledge_change_Addspledgechange(reals_dir_setup_cleanup):
-    # GIVEN
-    sue_text = "Sue"
-    sue_filehub = filehub_shop(None, None, sue_text)
+    sue_filehub = filehub_shop(env_dir(), root_label(), sue_text)
     sue_filehub.initialize_change_duty_files()
     old_sue_duty = sue_filehub.get_duty_agenda()
     clean_text = "clean"
@@ -53,10 +30,10 @@ def test_add_pledge_change_Addspledgechange(reals_dir_setup_cleanup):
     assert new_sue_duty.idea_exists(clean_road)
 
 
-def test_add_pledge_change_SetsDutyAgendapledgeIdea_suffgroup(reals_dir_setup_cleanup):
+def test_add_pledge_change_SetsDutyAgendapledgeIdea_suffgroup(env_dir_setup_cleanup):
     # GIVEN
     sue_text = "Sue"
-    sue_filehub = filehub_shop(None, None, sue_text)
+    sue_filehub = filehub_shop(env_dir(), root_label(), sue_text)
     sue_filehub.initialize_change_duty_files()
     old_sue_duty = sue_filehub.get_duty_agenda()
     clean_text = "clean"
