+++ src/real/test_real/test_real_work_agenda.py
@@ -1,53 +1,10 @@
 from src.agenda.healer import healerhold_shop
-from src.agenda.group import groupunit_shop
 from src.agenda.idea import ideaunit_shop
 from src.agenda.agenda import agendaunit_shop
 from src.change.filehub import filehub_shop
-from src.real.admin_work import get_default_work_agenda
 from src.real.real import realunit_shop
 from src.real.examples.real_env_kit import get_test_reals_dir, reals_dir_setup_cleanup
 from os.path import exists as os_path_exists
-
-
-def test_get_default_work_agenda_ReturnsCorrectObj():
-    # GIVEN
-    sue_text = "Sue"
-    blue_text = "blue"
-    slash_text = "/"
-    five_planck = 5
-    sue_party_pool = 800
-    casa_text = "casa"
-    bob_text = "Bob"
-    last_change_id = 7
-    sue_max_tree_traverse = 9
-    sue_agendaunit = agendaunit_shop(sue_text, blue_text, slash_text, five_planck)
-    sue_agendaunit.set_last_change_id(last_change_id)
-    sue_agendaunit.add_partyunit(bob_text, 3, 4)
-    swim_text = "/swimmers"
-    swim_groupunit = groupunit_shop(swim_text, _road_delimiter=slash_text)
-    swim_groupunit.edit_partylink(bob_text)
-    sue_agendaunit.set_groupunit(swim_groupunit)
-    sue_agendaunit.set_party_pool(sue_party_pool)
-    sue_agendaunit.add_l1_idea(ideaunit_shop(casa_text))
-    sue_agendaunit.set_max_tree_traverse(sue_max_tree_traverse)
-
-    # WHEN
-    default_work_agenda = get_default_work_agenda(sue_agendaunit)
-
-    # THEN
-    default_work_agenda.calc_agenda_metrics()
-    assert default_work_agenda._owner_id == sue_agendaunit._owner_id
-    assert default_work_agenda._owner_id == sue_text
-    assert default_work_agenda._real_id == sue_agendaunit._real_id
-    assert default_work_agenda._real_id == blue_text
-    assert default_work_agenda._road_delimiter == slash_text
-    assert default_work_agenda._planck == five_planck
-    assert default_work_agenda._party_creditor_pool is None
-    assert default_work_agenda._party_debtor_pool is None
-    assert default_work_agenda._max_tree_traverse == sue_max_tree_traverse
-    assert len(default_work_agenda.get_partys_dict()) == 1
-    assert len(default_work_agenda.get_groupunits_dict()) == 1
-    assert len(default_work_agenda._idea_dict) == 1
 
 
 def test_RealUnit_generate_work_agenda_Sets_work_AgendaFile(reals_dir_setup_cleanup):
